+++ raccoon/__init__.py
@@ -1,3 +1,3 @@
 
-__version__ = "1.1.9"
+__version__ = "1.2"
 from .dataframe import DataFrame, from_json